requirements.txt
pyserial
opencv-python
kivy=2.2.0
japanize-kivy=0.1.1
pyrealsense
pandas